Add data extraction category to fit policies
Following discussion around the monkeylearn submission, I've added this "data extraction" category to categories of packages that we accept.  It encompasses packages already created and accepted by us (`pdftools`, `tabulizer`) and fits in with our mission to enable data usage.  This expands the range of packages we explicitly accept without excessively broadening to "non-data APIs".